fix package and import files
