+++ Package.swift
@@ -1,11 +1,11 @@
-// swift-tools-version:5.3
+// swift-tools-version:5.7
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
 
 let package = Package(
     name: "CArch",
-    platforms: [.iOS(.v11)],
+    platforms: [.iOS(.v11), .macCatalyst(.v13)],
     products: [
         // Products define the executables and libraries a package produces, and make them visible to other packages.
         .library(
@@ -25,9 +25,9 @@ let package = Package(
             path: "Sources",
             exclude: ["Info.plist"]),
         .testTarget(
-            name: "carchTests",
+            name: "CArchTests",
             dependencies: ["CArch"],
-            path: "Tests",
+            path: "CArchTests",
             exclude: ["Info.plist"]),
     ]
 )
