+++ Sources/Ability/AlertAbility.swift
@@ -23,6 +23,7 @@
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 //  SOFTWARE.
 
+#if canImport(UIKit)
 import UIKit
 
 /// Протокол способности оповещения
@@ -171,3 +172,4 @@ public extension AlertAbility where Self: UIViewController {
         present(alert, animated: true)
     }
 }
+#endif
