+++ Sources/Ability/DocumentInteractionAbility.swift
@@ -23,6 +23,7 @@
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 //  SOFTWARE.
 
+#if canImport(UIKit)
 import UIKit
 
 /// Протокол работы с документами
@@ -87,3 +88,4 @@ public extension URL {
         return name
     }
 }
+#endif
