+++ Sources/Ability/KeyboardManagedAbility.swift
@@ -23,6 +23,7 @@
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 //  SOFTWARE.
 
+#if canImport(UIKit)
 import UIKit
 
 /// Тип слабая ссылка на объект
@@ -95,3 +96,4 @@ private extension UIViewController {
         view.endEditing(true)
     }
 }
+#endif
