+++ Sources/Ability/LoadingAbility.swift
@@ -23,6 +23,7 @@
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 //  SOFTWARE.
 
+#if canImport(UIKit)
 import UIKit
 
 /// Протокол получения доступа к заглушке прогресс загрузки
@@ -77,3 +78,4 @@ public extension LoadingAbility {
         hideLoading {}
     }
 }
+#endif
