+++ Sources/ModuleStructure/Routing.swift
@@ -23,6 +23,7 @@
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 //  SOFTWARE.
 
+#if canImport(UIKit)
 import UIKit
 
 /// Ошибка подготовки транзакции между модулями
@@ -85,3 +86,4 @@ public extension RootRoutingLogic {
         }
     }
 }
+#endif
