+++ Sources/Transition/Builders/ModuleBuilder.swift
@@ -23,7 +23,7 @@
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 //  SOFTWARE.
 
-import UIKit
+import Foundation
 
 /// Билдер любого модуля
 public protocol AnyModuleBuilder {
