+++ Sources/Transition/Builders/TransitionBuilder.swift
@@ -23,6 +23,7 @@
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 //  SOFTWARE.
 
+#if canImport(UIKit)
 import UIKit
 
 /// Билдер транзакции
@@ -153,3 +154,4 @@ public class TransitionBuilder {
         }
     }
 }
+#endif
