+++ Sources/Transition/Controller/TransitionController.swift
@@ -23,6 +23,7 @@
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 //  SOFTWARE.
 
+#if canImport(UIKit)
 import UIKit
 
 /// Замыкание завершения транзакции
@@ -424,3 +425,4 @@ private extension Array where Element == UIViewController {
         })
     }
 }
+#endif
