+++ Sources/Transition/SegueSettings/SplitTransitionControllerDelegate.swift
@@ -23,6 +23,7 @@
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 //  SOFTWARE.
 
+#if canImport(UIKit)
 import UIKit
 
 /// Протокол обозначения PrimaryViewController в SplitViewController
@@ -78,3 +79,4 @@ public final class SplitTransitionControllerDelegate: NSObject, UISplitViewContr
         }
     }
 }
+#endif
