+++ Sources/Transition/SegueSettings/UIStoryboardSegue+Identifier.swift
@@ -23,6 +23,7 @@
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 //  SOFTWARE.
 
+#if canImport(UIKit)
 import UIKit
 
 // MARK: - UIStoryboardSegue + CArch
@@ -60,3 +61,4 @@ extension UIStoryboardSegue {
         }
     }
 }
+#endif
