+++ Sources/Transition/SegueSettings/UIViewController+Name.swift
@@ -23,6 +23,7 @@
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 //  SOFTWARE.
 
+#if canImport(UIKit)
 import UIKit
 
 // MARK: - UIViewController + CArch
@@ -65,3 +66,4 @@ public extension UIViewController {
         }
     }
 }
+#endif
